+++ ididi/_node.py
@@ -132,9 +132,9 @@ def should_override(other_node: "DependentNode", current_node: "DependentNode") 
     Check if the other node should override the current node
     """
 
-    return (
-        ResolveOrder[other_node.factory_type] > ResolveOrder[current_node.factory_type]
-    )
+    other_priority = ResolveOrder[other_node.factory_type]
+    current_priority = ResolveOrder[current_node.factory_type]
+    return  other_priority > current_priority
 
 
 def resolve_marks(annt: Any) -> IDependent[Any]: