+++ ididi/errors.py
@@ -107,26 +107,7 @@ class UnsolvableParameterError(UnsolvableNodeError):
     Raised when a parameter is unsolveable.
     """
 
-
-# class UnsolvableDependencyError(UnsolvableParameterError):
-#     """
-#     Raised when a dependency parameter can't be built.
-#     """
-
-#     def __init__(
-#         self,
-#         *,
-#         dep_name: str,
-#         factory: Union[Callable[..., Any], type],
-#         dependent_type: Callable[..., Any],
-#         dependency_type: Callable[..., Any],
-#     ):
-#         type_repr = getattr(dependency_type, "__name__", str(dependency_type))
-#         param_repr = f" * {dependent_type.__name__}({dep_name}: {type_repr}) \n value of `{dep_name}` must be provided"
-#         self.message = (
-#             f"Unable to resolve dependency for parameter in {factory}, \n{param_repr}"
-#         )
-#         super().__init__(self.message)
+    
 
 
 class ForwardReferenceNotFoundError(UnsolvableParameterError):
@@ -224,3 +205,4 @@ class TopLevelBulitinTypeError(GraphResolveError):
         super().__init__(
             f"Using builtin type {dependency_type} as a top level dependency is not supported"
         )
+