+++ tests/test_node.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 import pytest
 
-from ididi import DependentNode
+from ididi import DependentNode, NodeConfig
 
 
 class A:
@@ -89,3 +89,15 @@ def test_node_from_annotated():
 
     node = DependentNode.from_node(hello)
     assert node.dependent is str
+
+
+def test_node_config_equality():
+    config1 = NodeConfig()
+    config2 = NodeConfig()
+
+    assert config1 == config2
+
+    
+    config3 = NodeConfig(reuse=True)
+
+    assert config1 != config3